ch4/generic: add framework for new long message protocol
Add two new handlers for PIPELINE and RDMA_READ protocol. The existing
LMT protocol is kept and will be removed in the following commits when
the PIPELINE and RDMA_READ are implemented.

Add origin and target callback functions for PIPELINE and RDMA_READ.
They are empty and waiting for implementation.

Add a new field in long_req msg hdr and long_ack msg hdr for the sender
and receiver to negotiate for the protocol of sending long message.  The
sender and receiver will negotiate for what protocol to use in RTS-CTS.
The send will send (in RTS) a list of bits for supported protocol based
on buffer location (CPU/GPU), packing, and is_local. The receiver would
choose one (after considering the buffer situation on the receive side)
from the received list and reply to the sender (in CTS). Then the sender
will start to use the corresponding interface in netmod/shmmod to send.
The receiver side will handle it through callback functions.

Add a new field in long_req msg hdr for the case when a AM msg uses AM
long message as carrier. This is for the use cases where a long SSEND
message or a long AM RMA message. This is to address the problem that
SSEND and RMA messages are ignoring the AM rndv patch and rely on the
netmod/shm's implicit support for long message.

Take SSEND as an example, the MPIDIG level SSEND implementation should
perform the send using MPIDIG_am_isend_impl which internally chooses
between AM eager or AM rndv for sending the actual data. If rndv is
chosen, the sender will prepare a long_req msg hdr with a payload
handler id set to MPIDIG_SSEND_REQ. The receiver side will perform
regular receiving of RTS and call corresponding target callback for the
payload handler id.  In this way, the following CTS and sending data can
be perform in the same way as a regular AM long message send.