ch4/generic: add framework for new long message protocol
Add new handlers for the AM messages in PIPELINE and RDMA_READ
protocols.  The existing LMT protocol is kept and will be removed in a
separate commits when the PIPELINE and RDMA_READ are implemented.

For the PIPELINE protocol, it has three new AM message types: RTS, CTS
and SEG. The RTS contains information for initiating the pipeline
protocol. The first segment of data will be piggbacked on the RTS. The
CTS is confirmation from receive side that the send has been matched and
ready for the rest segmets. The SEG is the transmission of the rest data
sengments.

For the RDMA_READ protocol, there are three new AM message types: REQ,
ACK and NAK. The REQ is for registering the send buffer with network and
sending the related info to receiver. The ACK is for the receiver to
reply that all the data has been received and the sender can safely
unregister and cleanup. The NAK is for the receiver to reject the
sender's attempt of doing RDMA_READ protocol and reply with a list of
preferred protocol for the tramission.

This commit adds the callback functions and message structs for these
new AM messages.

There is a new payload_handler_id field in MPIDIG_hdr_t for the case
when a AM msg uses these protocol as carrier. This is for the use cases
where a long SSEND message or a long AM RMA message. This is to address
the problem that SSEND and RMA messages relying on the netmod/shm's
implicit support for long message.

Take SSEND as an example, the MPIDIG level SSEND implementation should
perform the operation by sending the data through one of the EAGER,
PIPELINE and RDMA_READ protocol. In this case, the payload_handler_id
will be set as MPIDIG_SSEND_REQ and sending the actual data via the
protocol. The SSEND message header will be attached to the AM message
header of EAGER, PIPRLINE or RDMA_READ. The receiver side will read this
payload_handler_id and invoke the corresponding callback functions
to process the request correctly.