ch4/netmod: Add framework for new AM protocol.
Add interface for sending PIPELINE RTS. This is for the netmod to
initiate the PIPELINE protocol. It takes the parameters of the send
operation and prepare the netmod level send for RTS metadata and the
first segment of data.

Add interface for sending PIPELINE SEG. This is for sending the rest
data segment through the PIPELINE protocol.

Add interface for initiating RDMA_READ protocol. This interface takes
the send operation parameters, register send buffer with network and
sending the RDMA metadata to the receiver.