ch4/netmod: Add framework for new AM protocol.
Add interface for sending PIPELINE RTS. This is for the netmod to
initiate the PIPELINE protocol. It takes the parameters of the send
operation and prepare the netmod level send for RTS metadata and the
first segment of data.

Add interface for sending PIPELINE SEG. This is for sending the rest
data segment through the PIPELINE protocol.

Add interface for initiating RDMA_READ REQ. This interface takes
the send operation parameters, register send buffer with network and
sending the RDMA metadata to the receiver.

Add interface for starting RDMA_READ based receive on receiver side.

Add interface for unregister send side buffer after RDMA_READ is
completed.

Add new interface for checking protocol for AM buffer and OP. We have to
provide this because the AM RMA would run into the case where it need to
determine the protocol on target side. For example, the ACK of a GET and
the PUT where the target_addr's location can only be determined by the
target process. But because it is an RMA operation, because the target
process only have callbacks to handle the RMA, the regular checking
design in ofi_rma.h or posix_rma.h is insufficient---NM/SHM must expose
an interface for that.