ch4/shm: add framework for new AM long message protocol
Add interface for handling PIPELINE protocol send operation in SHM. This
is conceptually the same as the netmod interface.

Unlike netmod, the SHM only have PIPELINE as the long message send. It
does not have a RDMA_READ protocol. Note this could be extended in the
future when XPMEM wants to provided something similar to RDMA_READ (may
be with a different name).

A am_get_avail_long_protocol is added. This interface will take the
pointer to the send buffer, count and datatype as input and set the
availability of the long message protocol supported for the given send.
This gives additional long message protocol supported for the given send
beside PIPELINE. For SHM, this would be an empty function for now as
PIPELINE is the only protocol. But we could extend this in future.