ch4/posix: refactor POSIX PIPELINE
The main change of this patch is related to POSIX AM send path for long
message. Existing POSIX handles long message pipeline internally and
only use CH4 callback for completion. This commit make the following
changes.

1. Separating the eager path and pipeline path. This adopts a structure
for AM send that is similar to OFI. The old postpone queue structure is
removed and replaced by deferred queue structure. Any unused data
structure and fields are also removed.

2. Existing POSIX handles data copying in recv through calling
corresponding functions inside the progress engine. The target callback
was only used in preparing request and completion. This PR removes these
special treatment and relies on the callback to handle data movement for
receive. This unifies the recv path for AM long message from SHM and NM.
This also means when POSIX calls target_msg_cbs, the is_async is set to
zero. The corresponding special case for is_async == 1 in the callback
functions can be removed (not in this commit because of the other
dependecies).