ch4/generic: implement new PIPELINE protocol
For the sender, data_sz_left and offset are added to the request for
tracking the pipeline send progress.

Add a parent_req field for tracking potential related child request on
the sender side. When sending the buffer in multiple chunks, each of
these chunks may need a separate request object to track its on-the-fly
send operation. In this scenario, the original send request allocated at
MPIDIG_ level is the parent and will be used for sending the first
chunk. This parent request is also the one that is user visible. The
result chunks will need child requests allocated internally and
transparent to user. Each child request will set its parent_req field to
point at the parent_req, so that the sender side can correctly track and
complete them.

On the receive side, the callback function for MPIDIG_SEND_LONG_REQ
(RTS) will read the newly added field of available protocol bits and
prepare the request for segmented copy.

Befor receiver send out the CTS, it will choose a preferred protocol for
the data transmission. The default is PIPELINE. In the NM path, it will
check the receive buffer and decide whether it can use RDMA_READ
protocol.

When the sender receives the CTS, it will start data transmission using
the preferred protocol that the receiver selected.

The callback function for MPIDIG_SEND_LONG_PIPELINE copies/unpacks each
chunk into the receive buffer. It completes the request only when all
expected data is received. Note we have a special case for SHM here
because the SHM does copy internally and only invoke this callback when
it is done. This will be addressed in later commits.