ch4/ofi: implement PIPELINE am isend path and new EAGER am isend
Add MPIDI_OFI_am_send_request_t for OFI am send request. This separate
request data structure offers the benefit of creating multiple OFI level
am send request for one MPIR_Request in order to track the on-the-fly
send operations of each segment. Each of these request will contain its
own OFI level seq no for OFI order on receiver side. This request also
contains the pack buffer pointer so that we can release the pack buffer
at completion. All these OFI_am_send_requests are chained in a list in
the MPIR_Request.

Add MPIDI_OFI_do_am_isend_eager for sending AM over eager protocol. This
function largely copied the design of the existing do_am_isend function
with some changes. This function only send a piece of data that is
smaller than the eager threshold. This is to support both sending
short message and sending the PIPELINE RTS. In the long message case,
the first segment of data need to be piggybacked with the RTS, therefore
this new am_isend_eager will only send the a segment of data upto the
eager limit which becomes the first segment of the long message in
PIPELINE protocol.

Add MPIDIG_OFI_do_am_isend_pipeline for sending the long AM message
using pipeline protocol.  OFI will first check the datatype and buffer
location to determine if packing is needed.

If packing is needed, a pack buffer will be created which matches. A
segment of the user buffer will be packed into it. The combined size of
the actual packed data, OFI am header, CH4 am_hdr will be smaller than
the max short message size of OFI. If the packing is not needed, OFI
simply send from the user buffer segment by segment. The progress is
stored in the parent send request's data_sz_left and offset. These
parameters will be used to determine the src location for every send or
pack.

After sending the first segment, the OFI enqueues the rest operation as
a deferred isend and OFI progress will be responsible for issuing the
result operation.

If the send needs a pack buffer, but there is no free pack buffer
available, the operation will be deferred. If the deferred queue is not
empty, any new pipeline send operation is also deferred to maintain
order.

The progress engine will try issuing one segment at a time if a pack
buffer can be allocated. For OFI, each segment need to have a separate
request to track the on-the-fly send operation (this is because they
need to have different fi_context which is embedded inside the OFI
request object). Therefore, we may need to allocated ofi_am_send
requests for the issuing of deferred segments. However, if the segment
being issued is the first, we will use the original send request (parent
request) rather than creating an ofi_am_send request. This was due to
the cases in the previous paragraph. The management of ofi_am_send
request and parent request utilizes the parent_req field introduced in
the CH4 level.