ch4/posix: refactor EAGER interface.
This commit split the eager send interface into one that exposes the
payload buffer in the eager module and ont that performs the send op.

The eager_get_buf function allows the POSIX to get the payload of the
next available eager buffer in the eager module. This enables the POSIX
to pack data into the eager buffer directly.

The new interface og eager_send will take the eager buffer location and
perform the actual send operation. In iqueue, this is simply locating
the cell address using the eager buffer address and enqueue to the
target queue.