ch4/generic: Using MPIDIG_isend_imple for SSEND.
The existing SSEND message by passes the choice between eager send and
rndv send at CH4 level and relies on the SHM/NM's ability to send long
message. This commit changes the implementation to let the SSEND to use
the regular AM send message as carrier. This would allow the SSEND to
utilize the multi-protocol long message AM path that we added.

The implementation simply put the AM SSEMD msg inside the payload of a
regular AM SEND message. The handler id of MPIDIG_SSEND_REQ for the
SSEND is stored in the regular AM header (payload_handler_id). If the
payload_handler_id field of a header is not -1, it means this is not a
regular SEND but some other AM msg (in this case a SSEND msg).

The receive side will read if a payload_handler_id is not -1, the
target callback will invoke the corresponding callback function to
process the message. Note, in the case of SSEND, the SSEND target
callback is changed to only set the relavent field of the receive
request. The data copying process is still handled by the SEND target
callback and the long message callback for the chosen protocol.