ch4/generic: implement PIPELINE protocol at CH4 level.
Add new request type for pipeline send request. This contains the
information for PIPELINE send.

Add new field to MPIDIG_Request to tracking the progress of sending. It
has the data size that is waiting for send and the current offset.

Add one new entry function for invoking data transmission based on
protocols: EAGER, PIPELINE and RDMA_READ.

Add callback functions for the PIPELIEN protocol. The PIPELINE protocol
is initiated by sending the PIPELINE_RTS message along with the first
segment of data. The PIPELINE_RTS target callback will attemp the
matching if the incoming message with a posted receive request. If such
a posted receive is not available, it will copy the incoming data (first
segment) into buffer from unexp buffer pool and enqueue the message to
unexpected queue.

When the RTS is finally matched with a posted receive, it will send back
the PIPELINE_CTS message. Note, if the RTS was unexpected, we will need
to copy the first segment in the unexp buffer into the user receive
buffer before sending CTS.

When the sender receives the CTS, the CTS target callback will start the
rest data transmission by sending PIPELINE_SEG messages (one for each
segment). The new entry added in the request data structure is used for
tracking the progress.