ch4/generic: add framework for new long message protocol
Add new handlers for the AM messages in RDMA_READ
protocols.

For the RDMA_READ protocol, there are three new AM message types: REQ,
ACK and NAK. The REQ is for registering the send buffer with network and
sending the related info to receiver. The ACK is for the receiver to
reply that all the data has been received and the sender can safely
unregister and cleanup. The NAK is for the receiver to reject the
sender's attempt of doing RDMA_READ protocol and reply with a list of
preferred protocol for the tramission.