ch4/shm: Add framework for new AM protocol.
Add interface for sending PIPELINE RTS. This is for the netmod to
initiate the PIPELINE protocol. It takes the parameters of the send
operation and prepare the netmod level send for RTS metadata and the
first segment of data.

Add interface for sending PIPELINE SEG. This is for sending the rest
data segment through the PIPELINE protocol.