ch4/generic: add general request cache
The request cache is needed for the case of PIPELINE recv.
In PIPELINE send operation, the sender of only send the am_hdr on the
first segment. Therefore, the receiver will need to save the rreq
returned from the target callback for the subsequent segments. This
request cache functionality provides a mapping between a globally unique
key (for example, global rank of the src) and the rreq.