ch4/netmod: adding new interface for checking support for eager send
This interface takes am header size and the send buffer signature which
allows the netmod to do a more sophisticated checking. Right now, we
only check the size, we could be checking the buffer location and
determine if a ZCOPY transmission should be used.

The interface also allows passing the request which could be used to
save necessary information in the netmod section.