configure: create independent arch related options
The use of SSE2, AVX, or AVX512F instructions limits the build portability.
They should not be part of --enable-fast which handles general optimization
options.

`--enable-sse2` is default to on as most x86 CPUs supports it. If MPICH is
built for non-x86 architectures, configure will detect and disable it.
`--enable-avx` and `--enable-avx512f` is default to off for lack of universal
support.

These options check if the compiler supports `-msse2`, `-mavx` or `-mavx512f`.
If supported, configure will set CFLAGS accordingly. This can cause failure
if the MPICH is running on CPUs that do not support these instructions.